1) Fixed more bugs. 2) Tweaked code for clarity.
I previously forgot to update the list of entries with languages, kept
under the database key __ENTRIES_WITH_LANGUAGES__.

Also, the business of resetting old.languages instead of creating a
new temporary variable was stupid and would probably lead to confusion
and errors in some future update.  So for greater clarity, I used a
temporary variable.  Functionally it's no different than it was.